backend/requirements.txt
fastapi
uvicorn
sqlalchemy
aiomysql
aiofiles
reportlab
cryptography
google-genai
matplotlib
numpy
python-multipart